+++ src/stores/index.ts
@@ -1,4 +1,4 @@
 export { StoreProvider, useStore } from "./store-provider";
 export { RootStore, createRootStore, initRootStore } from "./root-store";
 export type { RootStoreHydration } from "./root-store";
-export { ExampleStore } from "./example-store";
+export { MachinesStore } from "./machines-store";